requirements.txt
matplotlib~=3.6.2
numpy~=1.23.4
mysql-connector-python~=8.0.31
adafruit-circuitpython-dht
